Dependency reorganization for cleanup, and to prepare for updates to Daemon
