+++ itsim/network/__init__.py
@@ -1,0 +1,9 @@
+from itsim import AbstractITObject
+
+
+class _Link(AbstractITObject):
+    pass
+
+
+class _Connection(AbstractITObject):
+    pass
