+++ itsim/node/process_management/process.py
@@ -1,5 +1,5 @@
 from itsim.simulator import Simulator
-from itsim.node import _Node
+from itsim.node.node_object import _Node
 from itsim.node.process_management import _Process
 from itsim.node.process_management.thread import Thread
 from itsim.types import Interrupt
