+++ itsim/node/router.py
@@ -1,5 +1,5 @@
 from itsim.network.link import Link
-from itsim.node import Node
+from itsim.node.node_object import Node
 
 
 class Router(Node):
